ADD: Update The way to see the texts was updated

+++ src/app/heroes/pages/add-heroes/add-heroes.component.ts
@@ -3,6 +3,9 @@ import { Heroe, Publisher } from '../../interfaces/heroes.interface';
 import { HeroesService } from '../../services/heroes.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs/operators';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmComponent } from '../../components/confirm/confirm.component';
 
 @Component({
   selector: 'app-add-heroes',
@@ -14,7 +17,9 @@ export class AddHeroesComponent implements OnInit {
   constructor(
     private readonly heroesService: HeroesService,
     private readonly activatedRoute: ActivatedRoute,
-    private readonly router: Router
+    private readonly router: Router,
+    private readonly _snackBar: MatSnackBar,
+    private readonly dialog: MatDialog
   ) { }
 
 
@@ -49,21 +54,41 @@ export class AddHeroesComponent implements OnInit {
 
     if(this.heroe.id) {
       this.heroesService.updateHero(this.heroe)
-      .subscribe(heroe => console.log(heroe));
+      .subscribe(heroe => {
+        console.log(heroe)
+        this.seeMessage('Hero Upgraded successfully')
+      });
     } else {
       this.heroesService.addHero(this.heroe)
       .subscribe(heroe => {
         console.log(heroe);
         this.router.navigate(['/heroes/edit-hero', heroe.id]);
+        this.seeMessage('Hero Created Successfully')
       });
     }
   }
 
   deleteHero() {
-    this.heroesService.deleteHero( this.heroe.id! )
-    .subscribe(res => {
-      console.log(res);
-      this.router.navigate(['/heroes'])
+     const dialog = this.dialog.open( ConfirmComponent, {
+      width: '250px',
+      data: { ...this.heroe }
+    })
+    dialog.afterClosed().subscribe(
+      (result) => {
+        if(result) {
+          this.heroesService.deleteHero( this.heroe.id! )
+          .subscribe(res => {
+            console.log(res);
+            this.router.navigate(['/heroes'])
+          })
+        }
+      }
+    )
+  }
+
+  seeMessage( msg: string ) {
+    this._snackBar.open( msg, 'Close', {
+      duration: 2500
     })
   }
 }
